[SPARK-53789][SQL][CONNECT] Unify error condition CANNOT_MODIFY_CONFIG
